Dedup clusters 4-10: consolidate small duplicated helpers into shared libs
- 4: parseCheckedBoxes + combineMapTags -> lib/map-field-utils.ts.
  map-update-logic and validate-publish used byte-identical copies;
  update-listing keeps a thin null-returning wrapper (verified equivalent for
  all inputs incl. '_No response_'); create-listing's parseTags reuses the
  shared checkbox parser and drops its combineMapTags copy.
- 5: new lib/time-buckets.ts with toHourBucketIso, toDateKey (YYYY-MM-DD) and
  toUtcDateKey (YYYY_MM_DD). discord-server-metrics/website-analytics re-export
  toDateKey+toHourBucketIso and railyard-app-downloads re-exports
  toHourBucketIso, keeping the public surface stable; railyard's local
  underscore toDateKey, backfill-core's toUtcDateKey and download-attribution's
  toUtcDateKeyFromIso all collapse into toUtcDateKey.
- 6: readListingIdsFromIndex(repoRoot, dir) -> lib/manifests.ts; converts
  sync-last-updated and sync-map-file-sizes (identical error strings).
  Kept local: map-demand-stats/repo.getMapIds (different error message) and
  regenerate-indexes' readdir-based getListingIds (no duplicate partner).
- 7: toWarningsOutputJson(prefix, warnings) -> lib/script-runtime.ts; the three
  copies differed only in prefix construction, now passed by callers verbatim.
- 8: listSnapshotFileNames + generic findLastAtOrBefore -> lib/history-utils.ts;
  analytics-core, ops/audit-download-history and railyard-app-downloads keep
  their own shapes/sorts on top of the truly shared parts.
- 9: parseRetryAfterMs -> lib/discord-webhook.ts (copies identical modulo
  quote style); capture-discord-server-metrics and make-announcement import it.
- 10: combineMapTags consolidated (see 4). Skipped as non-identical:
  sumDownloads (finite-guard vs unchecked add), fetchZipBuffer (different
  params/timeouts/warning text), parsePositiveInteger (null vs throwing),
  findDemandDataPath (try/catch stat vs existsSync+statSync).
