+++ scripts/announce/make-announcement.ts
@@ -1,6 +1,7 @@
 import fs from 'fs';
 import { pathToFileURL } from 'url';
 import { loadAuthorAliasIndex, resolveAuthorPresentation } from '../lib/author-aliases.js';
+import { parseRetryAfterMs } from '../lib/discord-webhook.js';
 import { resolveRepoRoot } from '../lib/script-runtime.js';
 
 const CONTENT_ANNOUNCEMENTS_ROLE_ID = '1492005223680446567';
@@ -23,20 +24,6 @@ const ALLOWED_MENTIONS = {
 
 function sleep(ms: number): Promise<void> {
     return new Promise((resolve) => setTimeout(resolve, ms));
-}
-
-function parseRetryAfterMs(payload: unknown): number | null {
-    if (!payload || typeof payload !== 'object') {
-        return null;
-    }
-    const retryAfter = (payload as { retry_after?: unknown }).retry_after;
-    if (typeof retryAfter !== 'number' || !Number.isFinite(retryAfter) || retryAfter < 0) {
-        return null;
-    }
-    if (retryAfter > 1000) {
-        return Math.ceil(retryAfter);
-    }
-    return Math.ceil(retryAfter * 1000);
 }
 
 async function sendAnnouncementRequest(webhookUrl: string, init: RequestInit): Promise<void> {
