+++ scripts/lib/map-field-utils.ts
@@ -21,6 +21,23 @@ export function isPresentIssueValue(value: unknown): value is string {
   return getOptionalIssueValue(value) !== undefined;
 }
 
+/** Selected labels from an issue-form checkbox field (array or markdown checkbox lines). */
+export function parseCheckedBoxes(raw: unknown): string[] {
+  if (Array.isArray(raw)) {
+    return raw.map((tag) => String(tag).trim()).filter(Boolean);
+  }
+  if (typeof raw !== "string" || !raw || raw === "_No response_") return [];
+  return raw
+    .split("\n")
+    .filter((line) => line.startsWith("- [X]") || line.startsWith("- [x]"))
+    .map((line) => line.replace(/^- \[[Xx]\]\s*/, "").trim())
+    .filter(Boolean);
+}
+
+export function combineMapTags(location: string, specialDemand: string[]): string[] {
+  return Array.from(new Set([location, ...specialDemand]));
+}
+
 export function getRequiredIssueValue(fieldName: string, value: unknown): string {
   const resolved = getOptionalIssueValue(value);
   if (!resolved) {
